requirements.txt
#streamlit>=1.32.0
mysql-connector-python>=8.3.0
pandas>=2.0.0
plotly>=5.18.0
flask >=2.0.0